+++ jwst/outlier_detection/outlier_detection.py
@@ -165,20 +165,13 @@ class OutlierDetection:
 
         # Perform median combination on set of drizzled mosaics
         median_model.data = self.create_median(drizzled_models)
-        if pars['save_intermediate_results']:
-            if self.outlierpars.get('asn_id', None) is None:
-                suffix_to_remove = self.resample_suffix
-            else:
-                suffix_to_remove = f"_{self.outlierpars['asn_id']}{self.resample_suffix}"
-            median_model_output_path = self.make_output_path(
-                basepath=median_model.meta.filename.replace(suffix_to_remove, '.fits'),
-                suffix='median')
-            median_model.save(median_model_output_path)
-            log.info(f"Saved model in {median_model_output_path}")
+
+        if self.outlierpars['save_intermediate_results']:
+            self.save_median(median_model) 
         else:
             # since we're not saving intermediate results if the drizzled models
             # were written to disk, remove them
-            if not pars['in_memory']:
+            if not self.outlierpars['in_memory']:
                 for fn in drizzled_models._models:
                     _remove_file(fn)
 
@@ -203,6 +196,7 @@ class OutlierDetection:
             for fn in blot_models._models:
                 _remove_file(fn)
         del median_model, blot_models
+    
 
     def create_median(self, resampled_models):
         """Create a median image from the singly resampled images.
@@ -218,35 +212,7 @@ class OutlierDetection:
 
         log.info("Computing median")
 
-        # Compute weight means without keeping datamodels for eacn input open
-        # Start by insuring that the ModelContainer does NOT open and keep each datamodel
-        ropen_orig = resampled_models._return_open
-        resampled_models._return_open = False  # turn off auto-opening of models
-        # keep track of resulting computation for each input resampled datamodel
-        weight_thresholds = []
-        # For each model, compute the bad-pixel threshold from the weight arrays
-        for resampled in resampled_models:
-            m = datamodel_open(resampled)
-            weight = m.wht
-            # necessary in order to assure that mask gets applied correctly
-            if hasattr(weight, '_mask'):
-                del weight._mask
-            mask_zero_weight = np.equal(weight, 0.)
-            mask_nans = np.isnan(weight)
-            # Combine the masks
-            weight_masked = np.ma.array(weight, mask=np.logical_or(
-                mask_zero_weight, mask_nans))
-            # Sigma-clip the unmasked data
-            weight_masked = sigma_clip(weight_masked, sigma=3, maxiters=5)
-            mean_weight = np.mean(weight_masked)
-            # Mask pixels where weight falls below maskpt percent
-            weight_threshold = mean_weight * maskpt
-            weight_thresholds.append(weight_threshold)
-            # close and delete the model, just to explicitly try to keep the memory as clean as possible
-            m.close()
-            del m
-        # Reset ModelContainer attribute to original value
-        resampled_models._return_open = ropen_orig
+        weight_thresholds = self.compute_weight_threshold(resampled_models, maskpt)
 
         # Now, set up buffered access to all input models
         resampled_models.set_buffer(1.0)  # Set buffer at 1Mb
@@ -282,6 +248,76 @@ class OutlierDetection:
             del resampled_sci, resampled_weight
 
         return median_image
+    
+
+    def save_median(self, median_model):
+        '''
+        Save median if requested by user
+
+        Parameters
+        ----------
+        median_model : ~jwst.datamodels.ImageModel
+            The median ImageModel or CubeModel to save
+        '''
+        
+        if self.outlierpars.get('asn_id', None) is None:
+            suffix_to_remove = self.resample_suffix
+        else:
+            suffix_to_remove = f"_{self.outlierpars['asn_id']}{self.resample_suffix}"
+        median_model_output_path = self.make_output_path(
+            basepath=median_model.meta.filename.replace(suffix_to_remove, '.fits'),
+            suffix='median')
+        median_model.save(median_model_output_path)
+        log.info(f"Saved model in {median_model_output_path}")
+
+
+    def compute_weight_threshold(self, resampled_models, maskpt):
+        '''
+        Compute weight means without keeping datamodels for each input open
+
+        Parameters
+        ----------
+        resampled_models : ~jwst.datamodels.ModelContainer
+            The input data models.
+
+        maskpt : float
+            The percentage of the mean weight to use as a threshold for masking.
+
+        Returns
+        -------
+        list
+            The weight thresholds for each integration.
+        '''
+
+        # Start by ensuring that the ModelContainer does NOT open and keep each datamodel
+        ropen_orig = resampled_models._return_open
+        resampled_models._return_open = True 
+        # keep track of resulting computation for each input resampled datamodel
+        weight_thresholds = []
+        # For each model, compute the bad-pixel threshold from the weight arrays
+        for resampled in resampled_models:
+            weight = resampled.wht
+            # necessary in order to assure that mask gets applied correctly
+            if hasattr(weight, '_mask'):
+                del weight._mask
+            mask_zero_weight = np.equal(weight, 0.)
+            mask_nans = np.isnan(weight)
+            # Combine the masks
+            weight_masked = np.ma.array(weight, mask=np.logical_or(
+                mask_zero_weight, mask_nans))
+            # Sigma-clip the unmasked data
+            weight_masked = sigma_clip(weight_masked, sigma=3, maxiters=5)
+            mean_weight = np.mean(weight_masked)
+            # Mask pixels where weight falls below maskpt percent
+            weight_threshold = mean_weight * maskpt
+            weight_thresholds.append(weight_threshold)
+            # close and delete the model, just to explicitly try to keep the memory as clean as possible
+            resampled.close()
+            del resampled
+        # Reset ModelContainer attribute to original value
+        resampled_models._return_open = ropen_orig
+        return weight_thresholds
+
 
     def blot_median(self, median_model):
         """Blot resampled median image back to the detector images."""