+++ jwst/outlier_detection/tests/test_outlier_detection.py
@@ -327,3 +327,44 @@ def test_outlier_step_image_weak_CR_nodither(exptype, tsovisit, tmp_cwd):
 
     # Verify CR is flagged
     assert result[0].dq[12, 12] == OUTLIER_DO_NOT_USE
+
+
+@pytest.mark.parametrize("exptype, tsovisit", exptypes_tso)
+def test_outlier_step_tso_rollingmed(exptype, tsovisit):
+    '''Test outlier detection with rolling median on time-varying source
+    This test fails if n_ints is set to 100, i.e., take simple median
+    '''
+    bkg = 1.5
+    sig = 0.02
+    numsci = 50
+    n_ints = 7 #rolling window size
+    signal = 7.0
+    container = ModelContainer(
+        we_many_sci(
+            numsci=numsci, background=bkg, sigma=sig, signal=signal, exptype=exptype, tsovisit=tsovisit
+        )
+    )
+
+    # Drop a weak CR on the science array
+    cr_timestep = 5
+    container[cr_timestep].data[12, 12] = bkg + sig * 10
+
+    # make time variability that has larger total amplitude than 
+    # the CR signal but deviations frame-by-frame are smaller
+    real_time_variability = signal * np.cos(np.linspace(0, np.pi, numsci))
+    for i, model in enumerate(container):
+        model.data[7,7] += real_time_variability[i]
+        model.err[7, 7] = np.sqrt(sig ** 2 + model.data[7,7])
+
+    result = OutlierDetectionStep.call(container, n_ints=n_ints)
+
+    # Make sure nothing changed in SCI array
+    for image, corrected in zip(container, result):
+        np.testing.assert_allclose(image.data, corrected.data)
+
+    # Verify source is not flagged
+    for r in result:
+        assert r.dq[7, 7] == datamodels.dqflags.pixel["GOOD"]
+
+    # Verify CR is flagged
+    assert result[cr_timestep].dq[12, 12] == OUTLIER_DO_NOT_USE